requirements.txt
# Core dependencies
nltk>=3.8.1
scikit-learn>=1.3.0
spacy>=3.7.2
yake>=0.4.8
transformers>=4.35.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2

# Optional dependencies for advanced features
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 